requirements.txt
python-telegram-bot[all]>=20.7,<22.0
anthropic>=0.36.0
groq>=0.9.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
httpx>=0.25.0
